+++ include/linux/kfence.h
@@ -1,4 +1,5 @@
 /* SPDX-License-Identifier: GPL-2.0 */
+
 #ifndef _LINUX_KFENCE_H
 #define _LINUX_KFENCE_H
